docker asses/requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
mysql-connector-python